requirements.txt
Django==2.0.7
django-cors-headers==2.3.0
Pillow==5.2.0
PyJWT==1.6.4
pyotp==2.2.6
pytz==2018.5
